Update blue_velvet.py
dictionaries were made into functions so they can read the variables accordingly (if they are just declared, the variables won't change). code is cleaner and the ID attribution is solved, but this causes a problem in calling function insertAudioTrack() because now I have to pass the idCounter var to it and everything else - solve by declaring global variable? or inserting declaration of attributes inside insertAudioTrack function?