Major update - createXML is now a function and created class ExportArdour for meny entry
To create a menu entry for the addon, Blender requires a class. To create this class, the core of the script had to be transformed into a function. Had to update all functions due to the loss of Global variables as a result to that. Also, corrected minor bug in function toSamples, using ar instead of audioRate as variable.