Fixed bug related to creation of sources and reinserted idCounter global variable
When you cut a strip in Blender, it names the resulting strip into .002, .003 etc instead of the original file extension .wav, .wov etc. This was causing some edited strips in the timeline to be directed to the "repeated" dictionaries list instead of the "sources" one. This is now being taken into consideration when creating the sources dictionary list. Status: fixed.

Also reinserted the variable idCounter as global, since the previous fix had inserted a bug in the creation of ids in the XML file. Status: fixed.