Major update in Space Squencer
This is a major change in Space Sequencer:

Strips infos of length, start and end in smpte are now in the "Strip data" section of the video monitoring;
The Timeline infos of timeline start and length are still in sequencer, but now isolated;
All SMPTE infos now show only MM:SS:FF if length is less than 59 min (59 min means it is close to 1h, so it's already best to show HH:MM:SS:FF);
A lot of the Sequencer properties that were redundant with "Strip data" panel that we see in video monitoring are now commented in the code, making these properties (the Sequencer) way more clear. This is yet experimental, but I'm quite confident the removed parts will leave no pounding hearts;
In "Strip data" panel (video monitoring), now there is also the option of changing strip source when the strip is an image. This was absent in the previous version, making the use of the sequencer properties necessary.