Fixed Strips Adjust scatterings
"Strips - Adjust to Cursor" and "Strips - Adjust to Start" now behave like we expect them to, fixing an odd scattering that happened when moving the strips around. This is a cosmetic change because both functions were fully functional, but it is always good to see exactly what you want to see happening in the timeline instead of some strips changing channels.