Blue Velvet ready for Blender 2.73a
Blue Velvet is now ready for Blender versions 2.73 and 2.73a. There's also an improvement on the way the addon reads the strips on the timeline - if user would have changed the strip's name to anything that didn't have an extension (such as audio.wav or audio_repeated.002), the script would throw the error "ValueError: need more than 1 value to unpack". This behaviour is now fixed. Thanks Bilka from France to pointing this out!