Ready for Blender 2.73a (fixed odd behaviour)
Blender comes with an odd "non-shortcut" for screen.screen_full_area by default - basically, if you press the mouse button on the screen, the screen will become fullscreen and hide all the buttons. Curiously, this does not seem to affect Blender when the set of shortcuts is set to default, but manages to screw up everything when you add other shortcuts such as the Velvet Shortcuts. So the code for this part is commented here for the sake of user sanity.